feat: add new components for avatar icon button label & switch
